+++ webapp/src/dog/dogClasses.ts
@@ -45,23 +45,4 @@ export const initLostDogWithPictureProps: ILostDogWithPicture = {
   isFound: false,
   picture:initPicture
 };
-export const initLostDogWithPictureProps2: ILostDogWithPicture = {
-  id: 1,
-  ownerId: 0,
-  pictureId: 0,
-  name: "Alex",
-  breed: "",
-  age: 0,
-  hairLength: "",
-  color: "",
-  size: "",
-  earsType: "",
-  tailLength: "",
-  specialMarks: "",
-  behaviors: [],
-  location: { city: "", district: "" },
-  dateLost: new Date(2012, 1, 1, 0, 0, 0, 0),
-  isFound: false,
-  picture:initPicture
-};
