+++ webapp/src/utilityComponents/uitilities.ts
@@ -1,0 +1,9 @@
+export type Item = {
+     filters: {
+        page: number;
+        size: number;
+    }, cookies:
+    {
+        [name: string]: any;
+    }
+}
